Refactor to prevent extra renders
