+++ react/hooks/useScrollOffset.ts
@@ -1,11 +1,11 @@
-import { useState, useLayoutEffect } from 'react'
+import { useState, useLayoutEffect, useCallback } from 'react'
 import throttle from 'throttleit'
 
 const listeners = new Set<Function>()
 
 const scrollHandler = throttle(() => {
   const scrollOffset = window.pageYOffset
-  listeners.forEach(set => set(scrollOffset))
+  listeners.forEach(fn => fn(scrollOffset))
 }, 100)
 
 function detach(fn: Function) {
@@ -28,8 +28,43 @@ function attach(fn: Function) {
  * It uses one single window scroll listener.
  * @returns {Number} - scroll position value
  */
-export const useScrollOffset = () => {
-  const [scroll, setScroll] = useState(0)
-  useLayoutEffect(() => attach(setScroll), [])
-  return { scroll }
+export const useScrollOffset = ({ offset }: { offset: number }) => {
+  // Initial element distance to the top of the container
+  const [initialOffsetTop, setInitialOffsetTop] = useState<number>()
+  const [hasReachedOffset, setReachedOffset] = useState(false)
+
+  // We use this callback as a ref to get the instance of the element once its mounted
+  // Reference: https://reactjs.org/docs/hooks-faq.html#how-can-i-measure-a-dom-node
+  const mountedDivRef = useCallback(
+    (node: HTMLDivElement) => {
+      if (node !== null) {
+        // `offsetTop` can be used for position:sticky, but not for position:fixed
+        // If we decide to change the css implementation for a js one, this should be revisited
+        setInitialOffsetTop(node.offsetTop)
+      }
+    },
+    // The rule below is disabled because we want to get a new callback every time
+    // the `offset` changes, which is a way to listen for possible sticky row size changes
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+    [offset]
+  )
+
+  useLayoutEffect(
+    () =>
+      attach((scroll: number) => {
+        const reached = initialOffsetTop && initialOffsetTop - scroll <= offset
+
+        if (!hasReachedOffset && reached) {
+          setReachedOffset(true)
+          return
+        }
+
+        if (hasReachedOffset && !reached) {
+          setReachedOffset(false)
+        }
+      }),
+    [hasReachedOffset, initialOffsetTop, offset]
+  )
+
+  return { hasReachedOffset, ref: mountedDivRef }
 }
